move components to coherent directories
