+++ src/components/layout/Layout.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components';
-import NavBar from '../components/Navbar'
-import Footer from '../components/Footer'
+import NavBar from './Navbar'
+import Footer from './Footer'
 import { Container } from 'semantic-ui-react'
 import 'semantic-ui-css/semantic.min.css'
