+++ src/pages/404.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Layout from '../components/Layout'
+import Layout from '../components/layout/Layout'
 
 const NotFoundPage = () => (
   <Layout>
