+++ src/pages/carriere/partners.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { graphql, StaticQuery } from 'gatsby'
 import Partner from '../../components/Partner';
-import ContentfulPage from '../../components/ContentfulPage'
+import ContentfulPage from '../../components/layout/ContentfulPage'
 import Markdown from 'markdown-to-jsx'
