+++ src/templates/PartnerTemplate.jsx
@@ -2,8 +2,8 @@ import React from 'react'
 import Markdown from 'markdown-to-jsx'
 import styled from 'styled-components'
 import { graphql } from 'gatsby'
-import Layout from '../components/Layout'
-import JobItem from '../components/JobItem'
+import Layout from '../components/layout/Layout'
+import JobItem from '../components/jobs/JobItem'
 
 const PartnerView = ({ data }) => {
   const { contentfulPartner: partner } = data
